Add optional token parameter
The token is used for querying GitHub Releases API to determine the
latest PMD version and the download URL for PMD.

This allows to use `@actions/github` to get a octokit.

Error handling is improved as well - util doesn't catch the errors anymore,
so any exceptions bubble up to the main function.