Log warning when MAX_PAGE is reached
This always logs a warning if we fetched page MAX_PAGE
regardless whether there would be more pages to fetch
or not. No effort is done to determine
whether the list of modified files is complete.

Also the modified files are now collected in
a Set to avoid duplicate entries.